Refactor shared code into a 'arguments' package
